requirements/local.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
mysqlclient==1.4.4
psycopg2-binary==2.8.4
PyMySQL==0.9.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
SQLAlchemy==1.3.12
urllib3==1.25.7
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
